Convert answer integration test
- These integration tests can be made into system tests.
- We create a helper for system tests using content items and begin moving the methods in from the old test_helper as they're used.
- for the moment these tests aren't including JS, so we can drive them with rack_test.
